Implement most of the plane of water.
This is a major upgrade from previous editions of this framework because the
code actually does something now. In particular, the plane of water can get
generated, and is managed through hooks rather than special code. You can now
safely revisit Water in wizmode, as a side effect.

The portal placement code hasn't been written yet, nor has the teleport region
code. The former should be straightforward, the latter is going to depend on how
I want to treat teleport regions and that's going to be tricky.