Purge in_mklev and improve mkobj() interface.
It started out as the former but the latter was necessary. Using flags for mkobj
and friends leads to a more consistent interface and will make it easier to add
more tweaks later.

in_mklev still exists, but it is set only by mklev() and now is used only for
impossible() checks in newsym() and vision_recalc(). They could be removed, but
I want to be sure that the behaviour is correct for some time before doing so.