Begin implementing a monster holding area.
migrating_mons is split into one chain for each level, which will hold all
monsters attempting to enter the level. Now, if a monster cannot enter the
level, it simply remains in the chain, waiting to enter again later. Also switch
steed dismounting to shove your steed onto this chain if it cannot be placed,
rather than killing it.

The plan is to eventually clean up all enexto-related stuff to usually move
monsters onto this waiting chain if we cannot find a place for them. This will
allow the arrival code to be cleaned up.